Improve explanation
Previously it was just: `<div>{message('Hello World')}</div>`

Which makes an impression that you could with that function, change the `className` of the surrounding div which is impossible. This makes it a bit more clear how the functions would work and that they have to return the entire `div`.