Align test spec for 07.04 with extra credit
As the spec is written, performing the extra credit will fail the
tests. That's because the spec expects certain form attributes, while
the extra credit says we can remove those attributes altogether.

One thing both of those conditions have in common (which the student
must change if they want the test to pass) is the form must no longer
GET from the endpoint.

While an argument for "progressive enhancement" _could_ be made here, I
agree with Kent's comment in the accompanying video that this is a
fully client-side app where passing the submission to the server doesn't
really make sense.

Alternatively, the "extra credit" comment could be updated to say
something like "FYI, the tests won't pass if you do this extra credit!"