Explain that `checked` can be a footgun in React
I ended up learning this on my own during the exercise, but now that
I've arrived at the "Default Value" lesson I'm seeing this is already
included in the course material.

Adding a call-out for it is kind of nice, especially since the
instructions start to talk about `checked` but don't explain the
read-only part.